another improvement

+++ keys.ts
@@ -9,7 +9,7 @@ import fs from 'fs';
     process.exit(1);
   }
 
-  console.log(`Generating ${amount} keys...`);
+  console.log(`\nGenerating ${amount} wallets...`);
 
   let keys: string[] = [];
   let wallets: {
@@ -28,6 +28,30 @@ import fs from 'fs';
     });
   }
 
-  fs.writeFileSync('./deps/generated-keys.txt', keys.join('\n'));
-  fs.writeFileSync('./deps/generated-wallets.json', JSON.stringify(wallets, null, 2));
+  if (
+    fs.existsSync('./deps/generated-keys.txt') ||
+    fs.existsSync('./deps/generated-wallets.json')
+  ) {
+    const parsedExisingKeys = fs
+      .readFileSync('./deps/generated-keys.txt', 'utf8')
+      .replaceAll('\r', '') // Remove carriage returns from Windows
+      .split('\n');
+    const exisingKeys = parsedExisingKeys.filter((key) => key !== '');
+
+    const parsedExistingWallets = fs.readFileSync('./deps/generated-wallets.json', 'utf8');
+    const existingWallets = parsedExistingWallets ? JSON.parse(parsedExistingWallets) : [];
+
+    if (exisingKeys.length > 0) {
+      console.log(`\nFound ${exisingKeys.length} existing wallets, merging...\n`);
+    }
+
+    exisingKeys.push(...keys);
+    existingWallets.push(...wallets);
+
+    fs.writeFileSync('./deps/generated-keys.txt', exisingKeys.join('\n'));
+    fs.writeFileSync('./deps/generated-wallets.json', JSON.stringify(existingWallets, null, 2));
+  } else {
+    fs.writeFileSync('./deps/generated-keys.txt', keys.join('\n'));
+    fs.writeFileSync('./deps/generated-wallets.json', JSON.stringify(wallets, null, 2));
+  }
 })();
